102-fibonacci.c

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -7,14 +7,14 @@
 
 int main(void)
 {
-	int n1 = 1, n2 = 2, n3, n;
+	long long n1 = 1, n2 = 2, n3, n;
 
-	printf("%d, %d", n1, n2);
+	printf("%lld, %lld", n1, n2);
 
 	for (n = 3; n <= 50; n++)
 	{
 		n3 = n1 + n2;
-		printf(", %d", n3);
+		printf(", %lld", n3);
 		n1 = n2;
 		n2 = n3;
 	}
